move enterprise/internal/cloud to internal/cloud
