+++ enterprise/cmd/frontend/internal/auth/sourcegraphoperator/BUILD.bazel
@@ -16,10 +16,10 @@ go_library(
         "//cmd/frontend/external/session",
         "//enterprise/cmd/frontend/internal/auth/openidconnect",
         "//enterprise/cmd/worker/shared/sourcegraphoperator",
-        "//enterprise/internal/cloud",
         "//internal/actor",
         "//internal/auth",
         "//internal/auth/providers",
+        "//internal/cloud",
         "//internal/conf",
         "//internal/conf/conftypes",
         "//internal/database",
@@ -48,10 +48,10 @@ go_test(
         "//cmd/frontend/external/session",
         "//enterprise/cmd/frontend/internal/auth/openidconnect",
         "//enterprise/cmd/worker/shared/sourcegraphoperator",
-        "//enterprise/internal/cloud",
         "//internal/actor",
         "//internal/auth",
         "//internal/auth/providers",
+        "//internal/cloud",
         "//internal/conf",
         "//internal/database",
         "//internal/database/dbtest",
