+++ enterprise/cmd/worker/internal/auth/BUILD.bazel
@@ -10,9 +10,9 @@ go_library(
         "//cmd/worker/job",
         "//cmd/worker/shared/init/db",
         "//enterprise/cmd/worker/shared/sourcegraphoperator",
-        "//enterprise/internal/cloud",
         "//internal/actor",
         "//internal/auth",
+        "//internal/cloud",
         "//internal/database",
         "//internal/env",
         "//internal/errcode",
@@ -34,8 +34,8 @@ go_test(
     ],
     deps = [
         "//enterprise/cmd/worker/shared/sourcegraphoperator",
-        "//enterprise/internal/cloud",
         "//internal/auth",
+        "//internal/cloud",
         "//internal/database",
         "//internal/database/dbtest",
         "//internal/extsvc",
